refactor: route を機能ごとの Sinatra クラスへ割る (#34)
app.rb の単一クラスから route を 5 つに割り、ClassLength の inline disable を
外した。App に残るのは合成だけ。

- lib/relay/base_app.rb … 設定 / DB / push クライアント / 共通 helper
- lib/relay/routes/{health,register,announcement_subscriptions,supporters,push}.rb

⚠ **束ね方は Sinatra の middleware（use）**。Rack::Cascade は status だけを見て
次のアプリへ渡すので、route が一致したうえでの halt 404（DELETE /register/:id の
「Not found」等）が後続に食われて別の応答に化ける。Sinatra は route 不一致の
ときだけ forward するのでこの事故が起きない。app_composition_test.rb で
「意図した 404 / 401 が次へ流れない」ことを route クラスごとに固定した。

設定は BaseApp の configure でクラス定義時に 1 度だけ組み立て、Sinatra の
settings 継承で各 route クラスへ渡る（DB も push クライアントも実体は 1 つ。
これも composition テストで固定）。

必須項目の検査は 3 route で同じ形だったので require_fields! に寄せた。
route 本体のロジックは動かしていない。

前コミットで入れた request テスト 49 件が分割の前後どちらでも緑であることを
確認済み（合計 116 runs）。
